Squashed commit of the following:
commit 96717173dbb69d8bc57cf494700632f9fcaa5c5d
Author: MrShadowek <77029131+MrShadow-X@users.noreply.github.com>
Date:   Tue Sep 10 19:14:40 2024 +0200

    update

commit fd3a31ced6e17f08d7dec174cd7df4befe006e3f
Author: MrShadowek <ciekawe.rzycie@gmail.com>
Date:   Fri Nov 3 16:24:35 2023 +0100

    Update - Fixing Bugs

    Fixed bugs, added new features.

commit 0e36bbe6e5c808570ee20c48d41715f5772f2bae
Author: MrShadowek <ciekawe.rzycie@gmail.com>
Date:   Sun Oct 29 12:28:34 2023 +0100

    .gitignore

commit 6d302b279e65d2e36c7e544f685229840847b7fb
Author: MrShadowek <ciekawe.rzycie@gmail.com>
Date:   Sun Oct 29 02:17:25 2023 +0200

    Small Fixes

    - Added {vPrefix}, {vSuffix} in rejoin message.
    - Countdown shortened message is now visible to everyone on arena.
    - Fixed void death message.
    - Fixed dropping bottles and buckets before removing them from inventory.

+++ bedwars-api/src/main/java/com/andrei1058/bedwars/api/arena/generator/IGenerator.java
@@ -77,7 +77,7 @@ public interface IGenerator {
     /**
      * Change item spawn delay. In seconds.
      */
-    void setDelay(int delay);
+    void setDelay(double delay);
 
     /**
      * Set how many items should the generator spawn at once.
@@ -136,16 +136,19 @@ public interface IGenerator {
 
     /**
      * Get the amount of items that are dropped once.
+     * @return
      */
     int getAmount();
 
     /**
      * Get spawn rate delay.
+     * @return
      */
-    int getDelay();
+    double getDelay();
 
     /**
      * Get seconds before next item spawn.
+     * @return
      */
     int getNextSpawn();
 
@@ -153,6 +156,7 @@ public interface IGenerator {
      * Get the spawn limit of the generators.
      * If there is this amount of items dropped near the generator
      * it will stop spawning new items.
+     * @return
      */
     int getSpawnLimit();
